Added try catch

+++ gems/core/views.py
@@ -381,24 +381,26 @@ def get_answer_values(request):
 
 def delete_contactgroup(request):
     if request.method == 'POST':
-        data = json.loads(request.body)
+        try:
+            data = json.loads(request.body)
+        except ValueError:
+            return HttpResponseBadRequest("Bad Request!")
 
-        if 'group_key' in data:
+        if 'group_key' in data and data['group_key'] != '':
             group_key = data['group_key']
 
-            group = ContactGroup.objects.filter(group_key=group_key).first()
-            api = ContactsApiClient(settings.VUMI_TOKEN)
-            deleted_group = api.delete_group(group_key)
+            try:
+                group = ContactGroup.objects.get(group_key=group_key)
+                api = ContactsApiClient(settings.VUMI_TOKEN)
+                deleted_group = api.delete_group(group_key)
+            except (Exception, ContactGroup.DoesNotExist):
+                return HttpResponseBadRequest('Bad Request!')
 
             if deleted_group['key'] == group_key:
                 group.delete()
-                return HttpResponse('OK')
-            else:
-                return HttpResponse('FAILED')
-        else:
-            return HttpResponse('FAILED')
+                return HttpResponse('Contact group deleted!')
 
-    return HttpResponse('BAD REQUEST TYPE')
+    return HttpResponseBadRequest('Bad Request!')
 
 
 def timing(f):
